requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.32.3
